requirements.txt
opencv-python==4.8.1.78
imutils==0.5.4
ultralytics==8.0.145
tensorflow==2.14.0
nats-py
minio
python-dotenv